+++ src/refactoring/features/admin/ui/index.ts
@@ -1,0 +1,4 @@
+export { NewProductForm } from './NewProductForm';
+export { ProductList } from './ProductList';
+export { CouponList } from './CouponList';
+export { NewCouponForm } from './NewCouponForm';
